requirements.txt
python-dotenv
uvloop
sanic
AoikLiveReload
aiomysql
sanic_compress
